Scanner now uses an external symbol table.
To save on space, symbol pointers are now stored in ident_t.meta instead of every symbol (including duplicates) being stored in ident_t.name

ident_t.meta points to an object in the symbol table, each of which is unique (duplicates will point to the same object).

Since ident_t.name is now unneeded, it was removed.